parse group definitions from gitolite.conf
* read_config parses group definitions from gitolite.conf
* this information is returned to the caller both as:
  - group memberships (which user belong to a group)
  - inverse group memberships (which groups is a user member of)

Note: in gitolite groups can be nested. This has not been tested so far,
thus nesting groups might lead to unexpected behavior.

Note: read_config takes the trac environment as input variable. this is
to be able to lug debug messages. this can be removed once the code is
stable.